backend/app/requirements.txt
FastAPI
uvicorn
sqlalchemy 
psycopg2-binary
python-dotenv
qrcode[pil]
reportlab